Refactoring and adding footer
