+++ packages/storefront/components/checkout/checkoutFooter.tsx
@@ -1,0 +1,21 @@
+import Link from "next/link";
+
+const CheckoutFooter = () => {
+  return (
+    <>
+      <div className="mt-10 flex flex-wrap gap-4">
+        <Link href="/" passHref>
+          <a className="text-decoration-none text-xs mt-2">Refund policy</a>
+        </Link>
+        <Link href="/" passHref>
+          <a className="text-decoration-none text-xs mt-2">Privacy policy</a>
+        </Link>
+        <Link href="/" passHref>
+          <a className="text-decoration-none text-xs mt-2">Terms of service</a>
+        </Link>
+      </div>
+    </>
+  );
+};
+
+export default CheckoutFooter;
