+++ packages/storefront/components/checkout/orderList/orderDetail.tsx
@@ -1,4 +1,3 @@
-import { NextComponentType } from "next";
 import React from "react";
 import CartProductList from "./cartProductList";
